mostly done - just need helper for smaller utils
