+++ eq-author-api/schema/resolvers/utils/routing.js
@@ -1,0 +1,59 @@
+const { filter, find, flatMap, some } = require("lodash");
+const { getPages } = require("./pages");
+const { getSkipConditions } = require("./skip");
+
+const getRouting = (ctx) =>
+  flatMap(filter(getPages(ctx), "routing"), "routing");
+
+const getRoutingById = (ctx, id) => find(getRouting(ctx), { id });
+
+const getRules = (ctx) => flatMap(filter(getRouting(ctx), "rules"), "rules");
+
+const getRoutingRuleById = (ctx, id) => find(getRules(ctx), { id });
+
+const getExpressionGroups = (ctx) =>
+  flatMap(filter(getRules(ctx), "expressionGroup"), "expressionGroup");
+
+const getAllExpressionGroups = (ctx) => {
+  const expressionGroups = getExpressionGroups(ctx);
+  return [...expressionGroups, ...getSkipConditions(ctx)];
+};
+
+const getExpressionGroupByExpressionId = (ctx, expressionId) =>
+  find(
+    getAllExpressionGroups(ctx),
+    (expressionGroup) =>
+      expressionGroup.expressions &&
+      some(expressionGroup.expressions, { id: expressionId })
+  );
+
+const getExpressionGroupById = (ctx, id) =>
+  find(getExpressionGroups(ctx), { id });
+
+const getExpressions = (ctx) => {
+  const routingExpressions = flatMap(
+    filter(getExpressionGroups(ctx), "expressions"),
+    "expressions"
+  );
+  const skipConditionExpressions = flatMap(
+    filter(getSkipConditions(ctx), "expressions"),
+    "expressions"
+  );
+
+  return [...routingExpressions, ...skipConditionExpressions];
+};
+
+const getExpressionById = (ctx, id) => find(getExpressions(ctx), { id });
+
+module.exports = {
+  getRouting,
+  getRoutingById,
+  getRules,
+  getRoutingRuleById,
+  getExpressionGroups,
+  getExpressionGroupById,
+  getExpressions,
+  getExpressionById,
+  getAllExpressionGroups,
+  getExpressionGroupByExpressionId,
+};
