+++ eq-author-api/schema/resolvers/utils/theme.js
@@ -1,0 +1,19 @@
+const { v4: uuidv4 } = require("uuid");
+
+const getThemeByShortName = ({ questionnaire }, shortName) =>
+  questionnaire.themes.find((theme) => theme.shortName === shortName);
+
+const createTheme = (attrs = {}) => ({
+  id: uuidv4(),
+  enabled: true,
+  shortName: "default",
+  legalBasisCode: "NOTICE_1",
+  eqId: "",
+  formType: "",
+  ...attrs,
+});
+
+module.exports = {
+  getThemeByShortName,
+  createTheme,
+};
